+++ src/Shaos.Services/ProcessInformation.cs
@@ -42,12 +42,12 @@ namespace Shaos.Services
         /// <summary>
         /// The maximum allowable working set size, in bytes, for the associated process.
         /// </summary>
-        public nint MaxWorkingSet { get; init; }
+        public long MaxWorkingSet { get; init; }
 
         /// <summary>
         /// The minimum allowable working set size, in bytes, for the associated process.
         /// </summary>
-        public nint MinWorkingSet { get; init; }
+        public long MinWorkingSet { get; init; }
 
         /// <summary>
         /// The amount of nonpaged system memory, in bytes, allocated for the associated process.
@@ -97,7 +97,7 @@ namespace Shaos.Services
         /// <summary>
         /// The processors on which the threads in this process can be scheduled to run.
         /// </summary>
-        public nint ProcessorAffinity { get; init; }
+        public long ProcessorAffinity { get; init; }
 
         /// <summary>
         /// The time that the associated process was started.
